+++ package.json
@@ -18,6 +18,8 @@
   },
   "homepage": "https://github.com/AdonisCreates/unitThree_backEnd#readme",
   "dependencies": {
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "querystring": "^0.2.0",
     "request": "^2.88.2",
